factor most code out of Jakefile into rhopm
 - Jakefile.contract -> rhopm.makeContractTask,
   deployContract
 - move shardIO, parseEnv to rhopm
 - startShard: never mind module.exports; just use task()
